fix: remove validateAlert.ts file instead use the already created alertvalidation file
